requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.3
requests==2.31.0
numpy==1.24.3
